resume_helper_backend/requirements.txt
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
anthropic==0.64.0
httpx==0.25.2
python-dotenv==1.0.0
gunicorn==21.0.0
openai>=1.12.0
python-docx==0.8.11
Werkzeug==2.3.7
boto3>=1.26.137
b2sdk>=1.24.0
PyPDF2>=3.0.1
reportlab>=4.0.4
pdfminer.six>=20240706
requests>=2.32.0

# LangChain packages
langchain>=0.1.0
langchain-core>=0.1.53
langchain-community>=0.0.38
langchain-openai>=0.1.5

# Document processing
docx2txt>=0.8

# System dependencies (install separately):
# LibreOffice: brew install --cask libreoffice (macOS) or sudo apt-get install libreoffice (Ubuntu)

# Additional dependencies
pydantic>=2.5.2
unstructured>=0.10.30
python-magic>=0.4.27
python-magic-bin>=0.4.14; sys_platform == 'win32'
typing-extensions>=4.5.0
tiktoken>=0.5.1
docx2pdf>=0.1.8
